+++ lib/Widgets/job_card.dart
