+++ lib/auth/main_page.dart
@@ -1,9 +1,10 @@
+
 import 'package:commiploy/pages/login_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import '../pages/home screens/home_page.dart';
+import '../Widgets/Menubar.dart';
 import 'auth_page.dart';
 
 class MainPage extends StatelessWidget {
@@ -15,7 +16,7 @@ class MainPage extends StatelessWidget {
       body: StreamBuilder<User?>(stream: FirebaseAuth.instance.authStateChanges(),
       builder: (context,snapshot){
         if(snapshot.hasData) {
-          return HomePage();
+          return Menubar();
         }else{
           return AuthPage();
         }
