+++ lib/pages/user screens/profile_page.dart
@@ -1,0 +1,42 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+
+class ProfilePage extends StatelessWidget {
+  final user = FirebaseAuth.instance.currentUser!;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.deepPurple[200],
+        elevation: 0,
+      ),
+
+    body: Column(
+    children: [
+      Center(
+      child: Container(
+        child: Text( user.email!,
+            style: TextStyle(fontSize: 50),
+          ),
+      ),
+    ),
+      Center(
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: GestureDetector(
+            onTap: () {
+              FirebaseAuth.instance.signOut();
+            },
+            child: Icon(Icons.logout),
+          ),
+        ),
+
+
+      ),
+    ],
+    )
+
+    );
+  }
+}
